% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_log_breaks.R
\name{calc_log_breaks}
\alias{calc_log_breaks}
\title{Gggplot2 is not very good at labeling the y-axis on log plots,
so always make sure we have at least one order of magnitude
on the scale, and 3 ticks.}
\usage{
calc_log_breaks(dat)
}
\arguments{
\item{dat}{y values}
}
\description{
Gggplot2 is not very good at labeling the y-axis on log plots,
so always make sure we have at least one order of magnitude
on the scale, and 3 ticks.
}
