% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mems_data.R
\name{create_mems_data}
\alias{create_mems_data}
\title{Create a matrix of MEMS data for use in TBsim from observed MEMS data
The function basically fills in any missing days when, and can
randomly draw more patients from the dataset.}
\usage{
create_mems_data(data = NULL, n_patients = NULL, n_events = 100,
  random = TRUE, seed = NULL, file = NULL)
}
\arguments{
\item{data}{Observed MEMS data (as list). See example.}

\item{n_patients}{number of patients to output}

\item{n_events}{number of days/dosing events per patient}

\item{random}{randomize the draw of patients?}

\item{seed}{randomization seed}

\item{file}{if not `NULL`, will write to specified CSV file}
}
\description{
Create a matrix of MEMS data for use in TBsim from observed MEMS data
The function basically fills in any missing days when, and can
randomly draw more patients from the dataset.
}
