% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_time_from_therapy.R
\name{max_time_from_therapy}
\alias{max_time_from_therapy}
\title{Get maximum time from therapy object}
\usage{
max_time_from_therapy(therapy)
}
\arguments{
\item{therapy}{object}
}
\description{
Get maximum time from therapy object
}

