% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{new_tempdir}
\alias{new_tempdir}
\title{Create new temp folder}
\usage{
new_tempdir(base = "/tmp/", len = 6)
}
\arguments{
\item{base}{base folder}

\item{len}{length of random character string}
}
\description{
Create new temp folder
}

