% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_saved_results.R
\name{read_saved_results}
\alias{read_saved_results}
\title{Read info from folder with saved runs}
\usage{
read_saved_results(folder = "/data/tbsim", user = "ronkeizer@gmail.com",
  read_all = FALSE, id = NULL)
}
\arguments{
\item{folder}{main folder where data is saved (default `/data`)}

\item{user}{username}

\item{read_all}{read all data (`TRUE``) or only header information (`FALSE`, default) from output data}
}
\description{
Read info from folder with saved runs
}

