% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_read_all_output.R
\name{tb_read_all_output}
\alias{tb_read_all_output}
\title{Read all output from TBsim}
\usage{
tb_read_all_output(folder, output_folder = FALSE, id = NULL, s3 = NULL,
  db = NULL)
}
\arguments{
\item{folder}{folder with output}

\item{db}{list with information to connect to MongoDB database}
}
\description{
Read all output from TBsim
}
