% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_read_headerfile.R
\name{tb_read_headerfile}
\alias{tb_read_headerfile}
\title{Created by John Fors, updates by Ron Keizer}
\usage{
tb_read_headerfile(folder = "~/tb_run")
}
\description{
Created by John Fors, updates by Ron Keizer
}
