% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb_read_output_db.R
\name{tb_read_output_db}
\alias{tb_read_output_db}
\title{Read output info from Db}
\usage{
tb_read_output_db(id, db = list())
}
\arguments{
\item{db}{list with information to connect to MongoDB database}
}
\description{
Read output info from Db
}

