% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBsim-package.R, R/TBsim.R
\docType{package}
\name{TBsim-package}
\alias{TBsim-package}
\title{TBsim package}
\description{
Wrapper around TBsim package

Wrapper around TBsim package
}
\author{
Ron Keizer \email{ronkeizer@gmail.com}

Ron Keizer \email{ronkeizer@gmail.com}
}

