% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{text_to_file}
\alias{text_to_file}
\title{Write lines of text to file}
\usage{
text_to_file(text, f)
}
\arguments{
\item{text}{vector of lines}

\item{f}{file name}
}
\description{
Write lines of text to file
}

